pyproject.toml
[project]
name = "kortreist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "geopandas>=1.1.1",
    "ipyleaflet>=0.20.0",
    "jupyterlab>=4.4.9",
    "openrouteservice>=2.3.3",
    "osmnx>=2.0.6",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ipykernel>=7.0.1",
]
